Restrict unsafe bidirectional pattern synonyms to unidirectional
We should be making the partial functions louder. The partial
bidirectional synonyms have dubious value since they give the
appearance of ordinary, total constructors. Thus these have been
downgraded to unidirectional only.

That said, writing 'unsafeFraction x y' is somewhat annoying, so we
_do_ add the partial infix function (%!). Hopefully the bang is clear
enough as to the danger.

To mitigate the danger, we also add an infix function for the total
mkFractionTH: (%%).