Swap prettyprinter for text-display
We were using prettyprinter for its Pretty typeclass, as we want two
"to string" options:

1. Represent the data as closely as possible (Show)
2. "Pretty" option for display purposes.

Bringing in a pretty printer library for 2 is overkill, so we use
the simpler text-display, which is designed exactly for this.

We also codify the following principle for 1:

We should derive Show as often as possible except when the type itself
contains extra information that is useful for debugging. Then we
write Show manually, adding in this extra type info. This is the case
for Interval, ModN, and ModP.

Accordingly, we change Fraction's Show instance from using the pattern
signature to its derived show instance.

We also remove Read instances, as Read is dubiously useful, and
complicates the Show story.