Add Internal modules exposing Unsafe constructors
It is occasionally useful to refer to internal constructors directly
(e.g. coerce) when we Know What We're Doing. These should generally
be exposed in an internal module with a scary name.

We also consistify error messages, consequently changing Fraction to
use error over Ratio's NonZero error (we additionally gain a
stacktrace by doing this).