Allow ModN/ModP to work over more types than just Integer/Natural
Previously we only allowed ModN/ModP to work with Integer/Natural.
This was to ensure that wrapping for a finite-type would not break
the modular arithmetic (e.g. consider Z/200Z for Int8).

To safely allow finite types, we implement size checks and perform
modular arithmetic in Integer before converting back to a, when needed.